backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-api-python-client
google-auth
reportlab
requests
pydantic